refactor: move messages and bridge modules
