+++ Dockerfile
@@ -45,7 +45,7 @@ RUN python3 -m pip install pip --upgrade && \
 # This significantly reduces the total build time, as each time we make a build call for a device,
 # only device-specific kernel source is being downloaded into the container.
 #
-RUN python3 ${WDIR}/wrapper/bridge.py --shared
+RUN python3 ${WDIR}/wrapper/utils/bridge.py --shared
 
 # launch app
 CMD [ "/bin/bash" ]
