+++ tests/unit/wrapper/tools/test_messages.py
@@ -1,6 +1,6 @@
 import pytest
 
-from wrapper.utils import messages as msg
+from wrapper.tools import messages as msg
 
 
 def test__message_note__validate(capfd):
