+++ wrapper/commands/bundle.py
@@ -5,8 +5,7 @@ import itertools
 from pydantic import BaseModel
 
 from wrapper.core import KernelBuilder, AssetsCollector
-from wrapper.tools import cleaning as cm, commands as ccmd, fileoperations as fo
-from wrapper.utils import messages as msg
+from wrapper.tools import cleaning as cm, commands as ccmd, fileoperations as fo, messages as msg
 from wrapper.configs import DirectoryConfig as dcfg
 from wrapper.interfaces import IBundleCommand
