+++ wrapper/configs/argument_config.py
@@ -4,8 +4,7 @@ from pathlib import Path
 from typing import Optional
 from pydantic import BaseModel
 
-from wrapper.tools import commands as ccmd
-from wrapper.utils import messages as msg
+from wrapper.tools import commands as ccmd, messages as msg
 
 
 class ArgumentConfig(BaseModel):
