+++ wrapper/core/assets_collector.py
@@ -1,8 +1,7 @@
 import os
 from pydantic import BaseModel
 
-from wrapper.tools import cleaning as cm, fileoperations as fo
-from wrapper.utils import messages as msg
+from wrapper.tools import cleaning as cm, fileoperations as fo, messages as msg
 from wrapper.clients import GitHubApi, LineageOsApi, ParanoidAndroidApi
 from wrapper.configs import DirectoryConfig as dcfg
 from wrapper.interfaces import IAssetsCollector
