+++ wrapper/core/kernel_builder.py
@@ -4,8 +4,8 @@ import time
 from pathlib import Path
 from pydantic import BaseModel
 
-from wrapper.tools import cleaning as cm, commands as ccmd, fileoperations as fo
-from wrapper.utils import ResourceManager, messages as msg
+from wrapper.tools import cleaning as cm, commands as ccmd, fileoperations as fo, messages as msg
+from wrapper.utils import ResourceManager
 from wrapper.configs import DirectoryConfig as dcfg
 from wrapper.interfaces import IKernelBuilder
