+++ wrapper/engines/docker_engine.py
@@ -1,8 +1,7 @@
 import os
 from typing import override
 
-from wrapper.tools import commands as ccmd
-from wrapper.utils import messages as msg
+from wrapper.tools import commands as ccmd, messages as msg
 from wrapper.interfaces import IDockerEngine
 
 from wrapper.engines.container_engine import ContainerEngine
