+++ wrapper/utils/resource_manager.py
@@ -3,8 +3,7 @@ import json
 import tarfile
 from typing import Optional
 
-from wrapper.tools import cleaning as cm, commands as ccmd, fileoperations as fo
-from wrapper.utils import messages as msg
+from wrapper.tools import cleaning as cm, commands as ccmd, fileoperations as fo, messages as msg
 from wrapper.configs import DirectoryConfig as dcfg
